pyproject.toml
[tool.poetry]
name = "hass-mcp-server"
version = "0.1.0"
description = "MCP Server for Home Assistant"
authors = ["ganhammar"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13.2"
mcp = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^25.9.0"
ruff = "^0.13.3"
pytest-homeassistant-custom-component = "^0.13.286"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = []
